re organize code structure of metartl, fix type annotations err
