+++ rllm/nn/encoder/ft_transformer_pre_encoder.py
@@ -4,7 +4,7 @@ from typing import Any, Dict, List
 from .table_pre_encoder import TablePreEncoder
 from .col_encoder._embedding_encoder import EmbeddingEncoder
 from .col_encoder._linear_encoder import LinearEncoder
-from rllm.types import ColType
+from rllm.types import ColType, StatType
 
 
 class FTTransformerPreEncoder(TablePreEncoder):
@@ -31,7 +31,7 @@ class FTTransformerPreEncoder(TablePreEncoder):
     def __init__(
         self,
         out_dim: int,
-        metadata: Dict[ColType, List[Dict[str, Any]]],
+        metadata: Dict[ColType, List[Dict[StatType, Any]]],
         in_dim: int = 1,
     ) -> None:
         col_encoder_dict = {
