+++ rllm/nn/models/metartl/metapath_prop_transform.py
@@ -1,17 +1,16 @@
-from typing import Dict, Iterable, Optional, Tuple
+from typing import Dict, Iterable, List, Optional, Tuple
 
 import torch
 from torch import Tensor
 
 from rllm.data import HeteroGraphData
-from rllm.transforms.utils import BaseTransform
 from rllm.transforms.utils.functional.metapath_prop import metapath_propagate
 
 
 EdgeType = Tuple[str, str, str]
 
 
-class MetaPathProp(BaseTransform):
+class MetaPathProp:
     r"""Generate multi-hop meta-path features for a target node type from a
     heterogeneous mini-batch, following MetaRTL
     `"MetaRTL: Meta-path Attention Enhanced Relational Table Learning"`.
@@ -30,13 +29,6 @@ class MetaPathProp(BaseTransform):
     Meta-path keys carry the full ``(src, rel, dst)`` triple of every
     traversed edge so parallel relations between the same pair of node types
     are not collapsed.
-
-    Note:
-        Unlike node- or edge-wise graph transforms, this operation consumes
-        both the graph structure and externally provided node features, and
-        returns a new feature tensor rather than a mutated graph. It therefore
-        lives under :mod:`rllm.transforms.utils` rather than the node/edge
-        graph transforms.
 
     Args:
         target_node_type (str): The node type to collect meta-path features for.
@@ -67,19 +59,11 @@ class MetaPathProp(BaseTransform):
         self.target_node_type = target_node_type
         self.min_hops = min_hops
         self.max_hops = max_hops
-        self.edge_schema = list(edge_schema) if edge_schema is not None else None
+        self.edge_schema: Optional[List[EdgeType]] = (
+            list(edge_schema) if edge_schema is not None else None
+        )
 
     def __call__(
-        self,
-        data: HeteroGraphData,
-        x_dict: Dict[str, Tensor],
-    ) -> Tensor:
-        # Meta-path propagation reads from the graph without mutating it, so we
-        # forward the data directly instead of shallow-copying as in the base
-        # class. The second ``x_dict`` argument carries the node features.
-        return self.forward(data, x_dict)
-
-    def forward(
         self,
         data: HeteroGraphData,
         x_dict: Dict[str, Tensor],
@@ -95,17 +79,15 @@ class MetaPathProp(BaseTransform):
             Tensor: Meta-path features of shape :obj:`[B, M, D]` ordered by
             meta-path name.
         """
-        # Do NOT drop empty edge types here: we need them to remain visible to
-        # `metapath_propagate` so it can zero-fill against `edge_schema` and
-        # keep the meta-path schema stable across batches.
-        edge_index_dict = dict(data.edge_index_dict)
-        num_nodes_dict = {
+        edge_index_dict: Dict[EdgeType, Tensor] = dict(data.edge_index_dict)
+        num_nodes_dict: Dict[str, int] = {
             node_type: node_store.num_nodes
             for node_type, node_store in data.node_items()
         }
 
         if self.edge_schema is None:
             self.edge_schema = list(edge_index_dict.keys())
+        schema: List[EdgeType] = self.edge_schema
 
         feats = metapath_propagate(
             edge_index_dict=edge_index_dict,
@@ -114,7 +96,7 @@ class MetaPathProp(BaseTransform):
             target_node_type=self.target_node_type,
             min_hops=self.min_hops,
             max_hops=self.max_hops,
-            edge_schema=self.edge_schema,
+            edge_schema=schema,
         )
         ordered_names = sorted(feats.keys())
         return torch.cat([feats[k] for k in ordered_names], dim=1)
