+++ rllm/transforms/table_transforms/tab_transformer_transform.py
@@ -2,7 +2,7 @@ from __future__ import annotations
 from typing import Any, Dict, List, Optional
 
 from rllm.transforms.table_transforms.col_normalize import ColNormalize
-from rllm.types import ColType
+from rllm.types import ColType, StatType
 from rllm.transforms.table_transforms import TableTransform
 
 
@@ -12,7 +12,7 @@ class TabTransformerTransform(TableTransform):
 
     Args:
         out_dim (int): The output dimensionality.
-        metadata (Dict[ColType, List[Dict[str, Any]]], optional): Metadata
+        metadata (Dict[ColType, List[Dict[StatType, Any]]], optional): Metadata
             containing information about the columns, such as statistics.
             (default: :obj:`None`)
     """
@@ -20,7 +20,7 @@ class TabTransformerTransform(TableTransform):
     def __init__(
         self,
         out_dim: int,
-        metadata: Optional[Dict[ColType, List[Dict[str, Any]]]] = None,
+        metadata: Optional[Dict[ColType, List[Dict[StatType, Any]]]] = None,
     ) -> None:
         super().__init__(
             out_dim=out_dim,
